+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 2.8.9.1
+
+ - Fix spell cycling via right-click (config option)
+
 ## 2.8.9
 
  - Fix cancelled batches (via undo) not being undoable after cancelling.